python3-homeassistant: relax pyyaml requirement
python3-pyyaml has been moved from meta-python to core, and in the
honister branch 5.4.1 is available.

Upstream homeassistant has used 5.4.1 for some releases (from 2021.2
to 2021.11), and I can't find anything that suggests that any changes
in the code were necessary; the pyyaml changelog from 5.3 to 5.4.1
also suggests it should be compatible.

Since bumping all of homeassistant to a newer release is a major
undertaking, let's start with this to see if it will work against
honister.